Adds two tests for backend that pass
EXTRA MOTIVATION NYANCAT TEST
