+++ src/server.js
@@ -1,21 +1,6 @@
 const http = require('http');
-const fs = require('fs');
-
+const handler = require('./src/handler.js');
 const port = process.env.PORT || 4000;
-
-function handler(req, res) {
-  const url = req.url;
-
-  if (url === '/') {
-    fs.readFile(__dirname + '/..' + '/index.html', (err, data) => {
-      if (err) {
-        throw err;
-      }
-      res.writeHead(200, {"Content-type": "text/html"});
-      res.end(data);
-    });
-  }
-}
 
 http.createServer(handler).listen(port);
